+++ VERSION_HISTORY.md
@@ -7,6 +7,7 @@ the top of the file.
     - Importing no longer uses/stores cached pickle files.
     - WorkflowManager.output() produces many table-format results if "count frequency" is False.
     - WorkflowManager.output() changes the column name for table-like outputs, according to the experiment.
+    - Greatly improved API documentation.
 * 2.2.1:
     - set music21 version to 2.1.2 exclusively
 * 2.2.0: